PracticaCompiladors/cl/requirements.txt
antlr4-python3-runtime==4.7.2
APScheduler==3.6.3
certifi==2020.12.5
cffi==1.14.4
cryptography==3.3.1
decorator==4.4.2
Pillow==8.1.0
pycparser==2.20
python-telegram-bot==13.1
pytz==2020.5
six==1.15.0
telegram==0.0.1
tornado==6.1
tzlocal==2.1
